--- src/main/java/com/arton/backend/auth/application/port/in/KaKaoUseCase.java
@@ -1,10 +1,0 @@
-package com.arton.backend.auth.application.port.in;
-
-import com.arton.backend.auth.application.data.OAuthSignupDto;
-import com.arton.backend.auth.application.data.TokenDto;
-
-import javax.servlet.http.HttpServletRequest;
-
-public interface KaKaoUseCase {
-    TokenDto login(HttpServletRequest request, OAuthSignupDto signupDto);
-}
