+++ src/main/java/com/arton/backend/auth/application/service/OAuthService.java
@@ -2,7 +2,7 @@ package com.arton.backend.auth.application.service;
 
 import com.arton.backend.auth.application.data.OAuthSignupDto;
 import com.arton.backend.auth.application.data.TokenDto;
-import com.arton.backend.auth.application.port.in.*;
+import com.arton.backend.auth.application.port.in.OAuthUseCase;
 import com.arton.backend.infra.shared.exception.CustomException;
 import com.arton.backend.infra.shared.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -10,27 +10,35 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
 public class OAuthService implements OAuthUseCase {
-    private final KaKaoUseCase kakaoService;
-    private final NaverUseCase naverService;
-    private final AppleUseCase appleService;
-    private final GoogleUseCase googleService;
+    private final AppleStrategy appleStrategy;
+    private final NaverStrategy naverStrategy;
+    private final KaKaoStrategy kaKaoStrategy;
+    private final GoogleStrategy googleStrategy;
+
+    private Map<Integer, OAuthStrategy> strategyMap = new HashMap<>()
+    {
+        {
+            put(0, kaKaoStrategy);
+            put(1, naverStrategy);
+            put(2, appleStrategy);
+            put(3, googleStrategy);
+
+        }
+    };
 
     @Override
     @Transactional
     public TokenDto signup(HttpServletRequest request, OAuthSignupDto signupDto) {
-        if (signupDto.getLoginType().equals("0")) {
-            return kakaoService.login(request, signupDto);
-        } else if (signupDto.getLoginType().equals("1")) {
-            return naverService.login(request, signupDto);
-        } else if (signupDto.getLoginType().equals("2")) {
-            return appleService.login(request, signupDto);
-        } else if (signupDto.getLoginType().equals("3")) {
-            return googleService.login(request, signupDto);
+        try {
+            return strategyMap.get(Integer.parseInt(signupDto.getLoginType())).signup(request, signupDto);
+        } catch (Exception e) {
+            throw new CustomException(ErrorCode.PARAMETER_NOT_VALID.getMessage(), ErrorCode.PARAMETER_NOT_VALID);
         }
-        throw new CustomException(ErrorCode.PARAMETER_NOT_VALID.getMessage(), ErrorCode.PARAMETER_NOT_VALID);
     }
 }
