+++ src/main/java/com/arton/backend/auth/application/service/OAuthStrategy.java
@@ -1,10 +1,10 @@
-package com.arton.backend.auth.application.port.in;
+package com.arton.backend.auth.application.service;
 
 import com.arton.backend.auth.application.data.OAuthSignupDto;
 import com.arton.backend.auth.application.data.TokenDto;
 
 import javax.servlet.http.HttpServletRequest;
 
-public interface GoogleUseCase {
-    TokenDto login(HttpServletRequest request, OAuthSignupDto signupDto);
+public interface OAuthStrategy {
+    TokenDto signup(HttpServletRequest request, OAuthSignupDto signupDto);
 }
